Add serial server and client.
Serial Server: TCP server handling arbitrary numbers of clients.
    Opens a serial device and pushes the input to all clients,
    Pushes input from any client to the serial output.

Serial Client: TCP client that pushes any TCP input to serial, and any
serial input to TCP.

This architecture allows many clients to send/recieve to a single
device, but not to each other (unless the server port is a loopback).